Safety check in

+++ src/main/java/grondag/adversity/niceblock/BigTexModelFactory.java
@@ -57,7 +57,7 @@ public class BigTexModelFactory extends ModelFactory
 	{
 		if (face == null) return QuadFactory.EMPTY_QUAD_LIST;
 
-        int clientShapeIndex = modelState.getClientShapeIndex(controller.getRenderLayer().ordinal());
+        int clientShapeIndex = (int) modelState.getClientShapeIndex(controller.getRenderLayer().ordinal());
         int faceIndex = FACE_SELECTORS[clientShapeIndex].selectors[face.ordinal()];
         int cacheKey = makeCacheKey(face, faceIndex, modelState.getColorIndex());
